sync chat from main work
