+++ public/sw.js
@@ -19,7 +19,6 @@ self.addEventListener('install', event => {
                 '/javascripts/addPlant.js',
                 '/javascripts/categories.js',
                 '/javascripts/changeUserProfile.js',
-                '/javascripts/viewPlant.js',
                 '/javascripts/idb-utility.js',
                 '/javascripts/imageUpload.js',
                 '/javascripts/index.js',
@@ -27,6 +26,7 @@ self.addEventListener('install', event => {
                 '/javascripts/overlay.js',
                 '/javascripts/plantUtilities.js',
                 '/javascripts/switchCategory.js',
+                '/javascripts/viewPlant.js',
                 '/stylesheets/main.scss',
                 '/stylesheets/main.css',
                 '/stylesheets/main.css.map',
